Add interface method to get human readable option names
Pull request #2
Co-authored-by: Priydarshi Singh <dryairship@gmail.com>

Print attributes/options and their choices are usually defined in a machine-readable form which is more made for easy typing in a command line, not too long, no special characters, always in English and human-readable form for GUI (print dialogs), more verbose for easier understanding, with spaces and other special characters, translated, ...

This commit adds functionality to the library to poll the human-readable strings from the backends.

Older backends without human-readable strings can still be used. In such a case it is recommended that the dialog does either its own conversion or simply shows the machine-readable string as a last mean.