Include all *.h files, they are all part of the API
The header files cpdb-libs-backend.h, cpdb-libs-frontend.h do not
contain the definitions of the API, they are simply convenience
header, including the headers needed for a backend and for a frontend
resp.

The actual headers are all in /usr/include/cpd-interface-headers, so
re-added the directory to the "make install" installation again.