Added test script for "make test"/"make check"
The script demo/run-tests.sh runs the demo/print_frontend text mode
example frontend and stops it by supplying "stop" to its standard
input.  If the frontend does not stop, it gets killed after a
timeout. The log of the frontend is checked whether it got actually
stopped by the "stop" command or not.

We cannot do actual backend discovery and print tests as we cannot
make the build of cpdb-libs dependent on any backend, due to the fact
that all backends depend on cpdb-libs. This circular dependency would
break bootstrapping (first complete build) a distro ...