Install sample frontend with "make install"
We use the sample frontend "print_frontend" for several tests now,
especially "make check" and also the autopkgtests in the Debian/Ubuntu
packages. They are also useful for backend developers for manual
testing.

Therefore we let it get installed now, so that it can be easily
included in distribution packages and also used by tests performed by
other packages, especially CPDB backends.

And ad print_frontend is now promoted to a general development/test
tool we renamed the demo/ directory to tools/.

We also discovered a segfault in the frontend library, in
cpdbResurrectPrinterFromFile(), when it is called with an invalid file
name, and fixed it.