Allow changing the backend info directory via env variable
To make it possible to test backends which are not installed into the
system, one can now set the environment variable CPDB_BACKEND_INFO_DIR
to the directory where the backend info file for the backend is, for
example in its source tree.

This especially allows to support "make check" for backends, which is
a requirement for them to get accepted in Ubuntu Main.