Handling saved settings: Make sure data structure is never NULL
cpdbReadSettingsFromDisk(): If there are no settings saved in a file
or the directory for the settings file cannot get created, no empty
data structure is created to hold any settings obtained while the
frontend is running, instead, simply NULL is returned which can lead
to crashes. What we need here is the empty data structure.

cpdbIgnoreLastSavedSettings(): To ignore the printer settings, the
function deletes the data structure and replaces it by NULL. Also here
we need an empty data structure instead.

Both these bugs are fixed now and empty data structures returned.