cpdb-text-frontend: Removed unnecessary g_main_loop
The text frontend spawned a sub-thread for running the command line
control and execution for it and in parallel, in the main thread, a
g_main_loop, to receive messages from the command execution, even if
the commands were executed asynchronously.

This is not needed, one can simply make the main thread waiting
without blocking any messages by just calling g_thread_join(). This
makes the code much simpler.

Moved the initial D-Bus connection and listing of available printers
also to the sub-thread, so that command execution starts only after
the printers having been listed.

As all the control goes into the sub-thread, renamed the function from
parse_commands() into control_thread().