Added support for CPDB backends running as permanent system daemon
The function cpdbActivateBackends() finds all available backends and
registers them with the frontend. For this it uses the
"ListActivatableNames" method which finds the backends by the
*.service files which are used to automatically start the program
which provides a given D-Bus services.

Now we want to package CPDB backends as Snap. When installing a Snap
we cannot install a *.service file, so we cannot make our backend
automatically started. Therefore the snapped backend we want to run as
a permanently running system daemon. The it registers itself on the
D-Bus and is available.

Problem is that the frontend does not search for already registered
services and so does not find this backend despite being readily
available, as there is no *.service file for it.

Now we use both the "ListActivatableNames" and the "ListNames"
methods, where the latter searches for already running services
regardless of *.service files. With both we always get a complete
list. We also check for duplicates, reported by both methods, now.